refactor(ui): git 分支胶囊收进变更侧栏 header 下方
- 左上角悬浮胶囊被用户判为突兀；分支与变更同域（仓库状态），
  收进 DiffSidebar：header 下方一条幽灵文字行（⎇ 分支名，11.5px
  ink-faint，hover 现 bg-hover 底），点击下弹只读分支列表（无边框
  浮卡 shadow-pop，当前分支圆点 + 加重），点击外部关闭
- 列表行无 hover/click 态避免假交互（切换仍走空态 ProjectBranch-
  Picker）；显示条件不变：有消息内容才显示、非 git 目录不渲染
- 分隔线从 .diff-side-head 移到 .diff-side-branch（head+branch
  同为一组头部）；数据仍走已有 IPC listGitBranches
- 删除消息区左上角悬浮胶囊 BranchBadge（App.tsx 挂载一并移除）
