+++ packages/desktop/src/renderer/src/styles/globals.css
@@ -979,13 +979,87 @@ code {
 	display: flex;
 	align-items: center;
 	gap: 10px;
-	padding: 12px 14px 10px;
-	border-bottom: 1px solid var(--color-border);
+	padding: 12px 14px 6px;
 }
 .diff-side-title {
 	font-size: 13px;
 	font-weight: 600;
 	color: var(--color-ink);
+}
+/* 分支行：header 下方幽灵文字行（⎇ 分支名），分隔线从 head 移到此处（head+branch 同为一组头部） */
+.diff-side-branch {
+	position: relative;
+	flex: none;
+	padding: 0 10px 8px;
+	border-bottom: 1px solid var(--color-border);
+}
+.diff-branch-btn {
+	display: inline-flex;
+	align-items: center;
+	gap: 6px;
+	max-width: 100%;
+	padding: 3px 6px;
+	border-radius: 6px;
+	font-size: 11.5px;
+	color: var(--color-ink-faint);
+	transition:
+		color 0.15s,
+		background-color 0.15s;
+}
+.diff-branch-btn:hover {
+	color: var(--color-ink-2);
+	background: var(--color-hover);
+}
+.diff-branch-name {
+	min-width: 0;
+	overflow: hidden;
+	text-overflow: ellipsis;
+	white-space: nowrap;
+}
+/* 只读分支列表下弹：无边框浮卡（shadow-pop），当前分支圆点 + 加重；行无 hover/click 态 */
+.diff-branch-pop {
+	position: absolute;
+	top: 100%;
+	left: 10px;
+	z-index: 30;
+	margin-top: 4px;
+	width: 240px;
+	max-height: 288px;
+	overflow-y: auto;
+	padding: 6px;
+	border-radius: 12px;
+	background: var(--color-surface);
+	box-shadow: var(--shadow-pop);
+}
+.diff-branch-item {
+	display: flex;
+	align-items: center;
+	gap: 8px;
+	padding: 5px 8px;
+	border-radius: 6px;
+	font-size: 13px;
+	color: var(--color-ink-2);
+}
+.diff-branch-item.on {
+	color: var(--color-ink);
+	font-weight: 500;
+}
+.diff-branch-dot {
+	width: 6px;
+	height: 6px;
+	flex: none;
+	border-radius: 9999px;
+	background: transparent;
+}
+.diff-branch-item.on .diff-branch-dot {
+	background: var(--color-ink-2);
+}
+.diff-branch-itemname,
+.diff-branch-item .diff-branch-itemname {
+	min-width: 0;
+	overflow: hidden;
+	text-overflow: ellipsis;
+	white-space: nowrap;
 }
 .diff-side-sum {
 	font-size: 11.5px;
